requirements.txt
# Core AI and API Dependencies
google-genai>=1.28.0
python-dotenv>=1.0.0

# Web Framework Dependencies
streamlit>=1.28.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Data Processing and Validation
pydantic>=2.5.0
requests>=2.31.0

# UI and Styling Dependencies
streamlit-option-menu>=0.3.6
streamlit-extras>=0.3.0

# Development and Testing (Optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0

# Additional Utilities
typing-extensions>=4.8.0
python-multipart>=0.0.6 